Removed view state and read data from webservice only for dates.
