+++ src/containers/control/add/index.js
@@ -22,7 +22,7 @@ import { connect } from 'react-redux'
 import { Grid, Row, Col, FormControl, PageHeader } from 'react-bootstrap'
 import PairedToogleButtons from '../../../components/control/add'
 import Field from '../../../components/formfields/musitfield'
-import { addControl } from '../../../reducers/control/add'
+import { addControl } from '../../../reducers/control'
 import Language from '../../../components/language'
 import DatePicker from 'react-bootstrap-date-picker'
 import moment from 'moment'
