+++ src/reducers/control/mapper/__tests__/mapper-test.js
