+++ src/reducers/control/mapper/templates/controlWebServiceGet.json
