+++ src/reducers/control/mapper/templates/controlsWithObservations.json
