+++ src/reducers/index.js
@@ -29,8 +29,7 @@ import storagePanelReducer from './storageunit/panel'
 import storageGridReducer from './storageunit/grid'
 import organizationReducer from './organization'
 import observationReducer from './observation'
-import controlReducer from './control/add'
-import controlDetailsReducer from './control/view'
+import controlReducer from './control'
 import observationControlGridReducer from './grid/observationcontrol'
 import nodeGridReducer from './grid/node'
 import objectGridReducer from './grid/object'
@@ -48,7 +47,6 @@ const rootReducer = combineReducers({
   organization: organizationReducer,
   observation: observationReducer,
   control: controlReducer,
-  controlDetails: controlDetailsReducer,
   observationControlGrid: observationControlGridReducer,
   nodeGrid: nodeGridReducer,
   objectGrid: objectGridReducer
